Stop using 3rd party players
Sadly, the Podverse player is not a viable option due to it preloading
part of the mp3. This breaks all playback statistics. Learning this lead
me to dig into using the html video tag and adding the track tag when I
have a WebVTT transcript available.

This also updates my site's headers